requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
requests-oauthlib==1.3.1
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
python-dotenv==1.0.0
gunicorn==21.2.0
